refactor: Provider를 providers 폴더로 분리
- components/providers/ 폴더 생성
- theme-provider.tsx 이동
- app/client-layout.tsx의 import 경로 업데이트
